refactor: extract CLI orchestration logic into focused modules
Extract large orchestration functions from cli.py and helper.py into
dedicated modules, making the codebase more maintainable and easier
to work with for both AI and humans.

Changes:
- Created seqerakit/core/block_parser.py: Extracted BlockParser class (88 lines)
  * Manages routing of YAML blocks to appropriate handlers
  * Handles on_exists behavior and resource deletion
  
- Created seqerakit/utils/file_discovery.py: Extracted find_yaml_files() function
  * Separates filesystem concerns from CLI
  * Supports files, directories, and stdin input
  
- Created seqerakit/core/yaml_processor.py: Extracted YAML processing orchestration
  * parse_all_yaml(): Loads and merges multiple YAML files
  * parse_yaml_block(): Parses individual resource blocks
  * parse_block(): Routes to resource-specific parsers
  * find_name(): Detects duplicate resources
  * RESOURCE_ORDER: Defines resource creation order
  
- Created seqerakit/resources/_deprecated.py: Moved deprecated handler wrappers
  * Makes legacy code obvious and contained
  * All functions emit DeprecationWarnings
  
- Updated seqerakit/cli.py: Simplified from 366 to ~200 lines
  * Imports from new modular locations
  * Focuses on CLI argument handling only
  
- Updated seqerakit/helper.py: Added explicit re-exports with __all__
  * Maintains backward compatibility for external code
  * Re-exports functions from new locations
  
- Updated tests to use correct module paths
  * All 111 tests passing

Benefits:
- Smaller, focused files with clear responsibilities
- Easier to understand and maintain individual modules
- Better separation of concerns (CLI, orchestration, file I/O)
- Consistent with modular pattern from PR1-PR3
- No breaking changes for end users
